Add files via upload
Upload of the powerpoint slidesI used at a presentation and discussion at the National Archvies, Kew, n July 19th 2024. The presentation was part of a series of presentations at the TNA to TNA staff at all levels as they began to explore the potential of LLMs for archival practive.

The presentattionw as titled: "A New Skill Set for Historians. Colin Greenstreet, TNA, Friday, July 19th 2024"